+++ CHANGELOG.md
@@ -12,6 +12,9 @@ even on minor releases, e.g. `v0.7.5` -> `v0.7.6`.
 
 *   Use real hyperlinks to issues in live view error alert.
     Fixes [#326](https://github.com/scottlamb/moonfire-nvr/issues/326).
+*   Fix 9x16 videos being stretched to the full screen dimensions in
+    list view full screen mode. Clean up some references to ancient
+    Firefox versions along the way.
 
 ## v0.7.26 (2026-02-17)